requirements.txt
scikit-learn
numpy
seaborn
matplotlib
tqdm
pillow
tk